# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ayo Indonesia Api",
        "description": "Ayo Indonesia Api Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/players": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "get list of players",
                "description": "Returns list of players",
                "operationId": "getPlayerList",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "create new player",
                "description": "create data player",
                "operationId": "createPlayer",
                "requestBody": {
                    "description": "create player",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Hansamu Yama"
                                    },
                                    "height": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "170"
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "70"
                                    },
                                    "number": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "5"
                                    },
                                    "team_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "4"
                                    },
                                    "positions": {
                                        "description": "player positions",
                                        "type": "array",
                                        "items": {
                                            "type": "number",
                                            "format": "number",
                                            "example": ""
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully Created"
                    },
                    "422": {
                        "description": "Unprocessable Content - Validation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/players/{id}": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "get player by id",
                "description": "get player by id",
                "operationId": "getPlayerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "get player by id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Players"
                ],
                "summary": "update player by id",
                "description": "update player data",
                "operationId": "updatePlayerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "update player by id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "update player by id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Hansamu Yama"
                                    },
                                    "height": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "170"
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "70"
                                    },
                                    "number": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "5"
                                    },
                                    "team_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "4"
                                    },
                                    "positions": {
                                        "description": "player positions",
                                        "type": "array",
                                        "items": {
                                            "type": "number",
                                            "format": "number",
                                            "example": ""
                                        },
                                        "example": [
                                            3,
                                            4
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content - Validation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Players"
                ],
                "summary": "delete player by id",
                "description": "delete player by Id",
                "operationId": "deletePlayerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "delete player by id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success - No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}